Move ProgressiveEnhancement into a separate file.
